change float -> Double in the _type_map.
The :class:`.Double` type is now used when a Python float value is detected
as a literal value to be sent as a bound parameter, rather than the
:class:`.Float` type.  :class:`.Double` has the same implementation as
:class:`.Float`, but when rendered in a CAST, produces ``DOUBLE`` or
``DOUBLE PRECISION`` rather than ``FLOAT``.   The former better matches
Python's ``float`` datatype which uses 8-byte double-precision storage.
Third party dialects which don't support the :class:`.Double` type directly
may need adjustment so that they render an appropriate keyword (e.g.
``FLOAT``) when the :class:`.Double` datatype is encountered.

Fixes: #10300

+++ test/sql/test_types.py
@@ -3699,7 +3699,7 @@ class ExpressionTest(
 
     @testing.combinations(
         (5, Integer),
-        (2.65, Float),
+        (2.65, Double),
         (True, Boolean),
         (decimal.Decimal("2.65"), Numeric),
         (datetime.date(2015, 7, 20), Date),
